Get rows and columns to decide on CRLF or not
Grab and store console.screen_rows and console.screen_columns. This 
helps with deciding if the tear line at the bottom needs a CRLF at 
the end before writing the credits, aka written/powered by.
Also, in the last commit I forgot to mention that we load 
utf8_cp437.js library to help take care of the UTF8 characters that
have accents, etc.